requirements.txt
annotated-types==0.7.0
anyio==4.10.0
babel==2.17.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
courlan==1.3.2
dateparser==1.2.2
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.7
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
tld==0.13.1
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.55.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
motor==3.7.1
pymongo==4.14.0
python-dotenv==1.0.0